gen-ai/rag-with-amazon-bedrock-knowledge-bases-using-s3-vectors/requirements.txt
bs4
requests
boto3 >= "1.39.14"
strands-agents
strands-agents-tools